requirements.txt
dvc==2.12.1
email_validator==1.2.1
fastds==0.6.0
flask_bootstrap==3.3.7.1
flask_cors==3.0.10
flask_login==0.6.1
flask_sqlalchemy==2.5.1
Flask_WTF==1.0.1
jupyter==1.0.0
matplotlib==3.5.2
mlflow==1.27.0
numpy==1.23.0
pandas==1.4.3
python_dotenv==0.20.0
PyYAML==6.0
